backend/requirements.txt
fastapi
uvicorn
pydantic
tortoise-orm
sse-starlette
requests
